added tests and force data to be float before checking for values

+++ signal_blocks/blocks/candle_close_block/events/close_eq_low.py
@@ -1,4 +1,4 @@
 def main(df, action):
     df["order"] = None
-    df.loc[df.close <= df.low, "order"] = action
+    df.loc[df.close.astype(float) <= df.low.astype(float), "order"] = action
     return df
